+++ changelog.md
@@ -2,3 +2,10 @@
 ## 2017.4.29
 - 增加php gd处理库的使用
 - https://github.com/Gregwar/Image/
+- 输出png图片
+
+````
+  ->get('png');
+header('Content-Type:image/png');
+echo $imgResult;
+```